Add files via upload
the two file paths (namely, strings 'input_filepath' and 'output_filepath') must be initialized before the code is run. the program reads the sentences line by line from the input file and for each of these sentences produces two POS and NER tagged output sentences, both of which are then written to the output file.